tutor factory

+++ database/migrations/2020_04_14_032327_create_students_table.php
@@ -18,7 +18,7 @@ class CreateStudentsTable extends Migration
             $table->string('student_name', 100);
             $table->string('student_email', 20);
             $table->string('student_phone', 100);
-            $table->unsignedInteger('tutor_ID');
+            $table->unsignedBigInteger('tutor_ID');
             $table->timestamps();
         });
     }
